requirements.txt
customtkinter>=5.2.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
python-docx>=1.0.0
